[BEBIG-71] chore : fin-type / fin-test 이동 UserController -> HomeController
