+++ src/main/java/beBig/mapper/UserMapper.java
@@ -31,8 +31,4 @@ public interface UserMapper {
     List<UtilVo> getUtilTerms(); // 약관 조회 메서드
 
     Long getUserIdByKaKaoId(String userLoginId);
-
-    List<FinTestVo> findFinTest();
-
-    FinTypeVo findFinTypeByUserId(long userId);
 }
