+++ src/main/java/beBig/service/UserServiceImp.java
@@ -183,32 +183,6 @@ public class UserServiceImp implements UserService {
         return userMapper.getUserIdByKaKaoId(kakaoId);
     }
 
-    @Override
-    public List<FinTestVo> findMission() {
-        UserMapper userMapper = sqlSessionTemplate.getMapper(UserMapper.class);
-        List<FinTestVo> list = userMapper.findFinTest();
-        return list;
-    }
-
-    @Override
-    public FinInfoResponseDto findFinTypeByUserId(Long userId) {
-        UserMapper userMapper = sqlSessionTemplate.getMapper(UserMapper.class);
-        FinTypeVo vo = userMapper.findFinTypeByUserId(userId);
-        FinInfoResponseDto type = new FinInfoResponseDto();
-
-        type.setFinTypeAnimal(vo.getFinTypeAnimal());
-        type.setFinTypeTitle(vo.getFinTypeTitle());
-        type.setFinTypeHabit1(vo.getFinTypeHabit1());
-        type.setFinTypeHabit2(vo.getFinTypeHabit2());
-        type.setFinTypeAnimalDescription(vo.getFinTypeAnimalDescription());
-        type.setFinTypeTitleDescription(vo.getFinTypeTitleDescription());
-        type.setFinTypeCode(vo.getFinTypeCode());
-
-        log.info("type : {}", type);
-
-        return type;
-    }
-
 
     //    public UserVo findByUserId(String userId) throws Exception {
 //        return userMapper.findByUserId(userId);
