+++ src/main/resources/beBig/mapper/HomeMapper.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<!DOCTYPE mapper
+        PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
+        "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
+
+<mapper namespace="beBig.mapper.HomeMapper">
+    <select id="findFinTest" resultType="beBig.vo.FinTestVo">
+        SELECT *
+        FROM fin_test;
+    </select>
+    <select id="findFinTypeByUserId" resultType="beBig.vo.FinTypeVo" parameterType="long">
+        SELECT f.*
+        FROM fin_type f
+                 JOIN user u
+                      ON u.fin_type_code = f.fin_type_code
+        WHERE u.user_id = #{userId};
+    </select>
+</mapper>
